fix: adds encoding option when opening the file
Also adds to scan file the decimal precision option

Fixes the type of argument when reading the file at the scan command.

Was causing a difference between size results from the size command and scan command, because at the scan function, when not reading the stdin, the filesize function was receiving the file content as str, not path. Now the size at the scan command follows the single size command.